0.2.1 Logger

+++ backend/internal/service/space.go
@@ -6,6 +6,7 @@ import (
 	"nora/internal/model"
 
 	"github.com/google/uuid"
+	"go.uber.org/zap"
 )
 
 func (s *Service) CreateSpace(ctx context.Context, data *model.SpaceCreate) (*model.Space, error) {
@@ -13,6 +14,10 @@ func (s *Service) CreateSpace(ctx context.Context, data *model.SpaceCreate) (*mo
 	if err != nil {
 		return nil, err
 	}
+
+	logger := ctx.Value("logger").(*zap.Logger)
+	logger.Info("created space", zap.String("id", id.String()))
+
 	return s.sp.FindOne(ctx, *id)
 }
 
@@ -39,8 +44,13 @@ func (s *Service) DeleteSpace(ctx context.Context, id uuid.UUID, userID uuid.UUI
 	if err != nil {
 		return err
 	}
+	if space == nil {
+		return e.ErrNotFound
+	}
 	if space.UserID != userID {
 		return e.ErrForbidden
 	}
+	logger := ctx.Value("logger").(*zap.Logger)
+	logger.Info("space deleted", zap.String("id", id.String()))
 	return s.sp.Delete(ctx, id)
 }
